@@ -1,14 +1,18 @@
 module github.com/bakito/sealed-secrets-web
 
-go 1.20
+go 1.21
 
 require (
+	github.com/bakito/semver v1.1.3
 	github.com/bitnami-labs/sealed-secrets v0.24.5
 	github.com/gin-gonic/gin v1.9.1
-	github.com/golang/mock v1.6.0
+	github.com/golangci/golangci-lint v1.55.2
 	github.com/google/uuid v1.5.0
+	github.com/goreleaser/goreleaser v1.23.0
+	github.com/norwoodj/helm-docs v1.12.0
 	github.com/onsi/ginkgo/v2 v2.13.2
 	github.com/onsi/gomega v1.30.0
+	go.uber.org/mock v0.4.0
 	gopkg.in/yaml.v3 v3.0.1
 	k8s.io/api v0.28.4
 	k8s.io/apimachinery v0.28.4
