@@ -1,5 +1,10 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: k8s.io/client-go/kubernetes/typed/core/v1 (interfaces: CoreV1Interface,SecretInterface)
+//
+// Generated by this command:
+//
+//	mockgen -destination pkg/mocks/core/mock.go --package core k8s.io/client-go/kubernetes/typed/core/v1 CoreV1Interface,SecretInterface
+//
 
 // Package core is a generated GoMock package.
 package core
@@ -8,7 +13,7 @@ import (
 	context "context"
 	reflect "reflect"
 
-	gomock "github.com/golang/mock/gomock"
+	gomock "go.uber.org/mock/gomock"
 	v1 "k8s.io/api/core/v1"
 	v10 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	types "k8s.io/apimachinery/pkg/types"
@@ -64,7 +69,7 @@ func (m *MockCoreV1Interface) ConfigMaps(arg0 string) v12.ConfigMapInterface {
 }
 
 // ConfigMaps indicates an expected call of ConfigMaps.
-func (mr *MockCoreV1InterfaceMockRecorder) ConfigMaps(arg0 interface{}) *gomock.Call {
+func (mr *MockCoreV1InterfaceMockRecorder) ConfigMaps(arg0 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ConfigMaps", reflect.TypeOf((*MockCoreV1Interface)(nil).ConfigMaps), arg0)
 }
@@ -78,7 +83,7 @@ func (m *MockCoreV1Interface) Endpoints(arg0 string) v12.EndpointsInterface {
 }
 
 // Endpoints indicates an expected call of Endpoints.
-func (mr *MockCoreV1InterfaceMockRecorder) Endpoints(arg0 interface{}) *gomock.Call {
+func (mr *MockCoreV1InterfaceMockRecorder) Endpoints(arg0 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Endpoints", reflect.TypeOf((*MockCoreV1Interface)(nil).Endpoints), arg0)
 }
@@ -92,7 +97,7 @@ func (m *MockCoreV1Interface) Events(arg0 string) v12.EventInterface {
 }
 
 // Events indicates an expected call of Events.
-func (mr *MockCoreV1InterfaceMockRecorder) Events(arg0 interface{}) *gomock.Call {
+func (mr *MockCoreV1InterfaceMockRecorder) Events(arg0 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Events", reflect.TypeOf((*MockCoreV1Interface)(nil).Events), arg0)
 }
@@ -106,7 +111,7 @@ func (m *MockCoreV1Interface) LimitRanges(arg0 string) v12.LimitRangeInterface {
 }
 
 // LimitRanges indicates an expected call of LimitRanges.
-func (mr *MockCoreV1InterfaceMockRecorder) LimitRanges(arg0 interface{}) *gomock.Call {
+func (mr *MockCoreV1InterfaceMockRecorder) LimitRanges(arg0 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "LimitRanges", reflect.TypeOf((*MockCoreV1Interface)(nil).LimitRanges), arg0)
 }
@@ -148,7 +153,7 @@ func (m *MockCoreV1Interface) PersistentVolumeClaims(arg0 string) v12.Persistent
 }
 
 // PersistentVolumeClaims indicates an expected call of PersistentVolumeClaims.
-func (mr *MockCoreV1InterfaceMockRecorder) PersistentVolumeClaims(arg0 interface{}) *gomock.Call {
+func (mr *MockCoreV1InterfaceMockRecorder) PersistentVolumeClaims(arg0 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "PersistentVolumeClaims", reflect.TypeOf((*MockCoreV1Interface)(nil).PersistentVolumeClaims), arg0)
 }
@@ -176,7 +181,7 @@ func (m *MockCoreV1Interface) PodTemplates(arg0 string) v12.PodTemplateInterface
 }
 
 // PodTemplates indicates an expected call of PodTemplates.
-func (mr *MockCoreV1InterfaceMockRecorder) PodTemplates(arg0 interface{}) *gomock.Call {
+func (mr *MockCoreV1InterfaceMockRecorder) PodTemplates(arg0 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "PodTemplates", reflect.TypeOf((*MockCoreV1Interface)(nil).PodTemplates), arg0)
 }
@@ -190,7 +195,7 @@ func (m *MockCoreV1Interface) Pods(arg0 string) v12.PodInterface {
 }
 
 // Pods indicates an expected call of Pods.
-func (mr *MockCoreV1InterfaceMockRecorder) Pods(arg0 interface{}) *gomock.Call {
+func (mr *MockCoreV1InterfaceMockRecorder) Pods(arg0 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Pods", reflect.TypeOf((*MockCoreV1Interface)(nil).Pods), arg0)
 }
@@ -218,7 +223,7 @@ func (m *MockCoreV1Interface) ReplicationControllers(arg0 string) v12.Replicatio
 }
 
 // ReplicationControllers indicates an expected call of ReplicationControllers.
-func (mr *MockCoreV1InterfaceMockRecorder) ReplicationControllers(arg0 interface{}) *gomock.Call {
+func (mr *MockCoreV1InterfaceMockRecorder) ReplicationControllers(arg0 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ReplicationControllers", reflect.TypeOf((*MockCoreV1Interface)(nil).ReplicationControllers), arg0)
 }
@@ -232,7 +237,7 @@ func (m *MockCoreV1Interface) ResourceQuotas(arg0 string) v12.ResourceQuotaInter
 }
 
 // ResourceQuotas indicates an expected call of ResourceQuotas.
-func (mr *MockCoreV1InterfaceMockRecorder) ResourceQuotas(arg0 interface{}) *gomock.Call {
+func (mr *MockCoreV1InterfaceMockRecorder) ResourceQuotas(arg0 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ResourceQuotas", reflect.TypeOf((*MockCoreV1Interface)(nil).ResourceQuotas), arg0)
 }
@@ -246,7 +251,7 @@ func (m *MockCoreV1Interface) Secrets(arg0 string) v12.SecretInterface {
 }
 
 // Secrets indicates an expected call of Secrets.
-func (mr *MockCoreV1InterfaceMockRecorder) Secrets(arg0 interface{}) *gomock.Call {
+func (mr *MockCoreV1InterfaceMockRecorder) Secrets(arg0 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Secrets", reflect.TypeOf((*MockCoreV1Interface)(nil).Secrets), arg0)
 }
@@ -260,7 +265,7 @@ func (m *MockCoreV1Interface) ServiceAccounts(arg0 string) v12.ServiceAccountInt
 }
 
 // ServiceAccounts indicates an expected call of ServiceAccounts.
-func (mr *MockCoreV1InterfaceMockRecorder) ServiceAccounts(arg0 interface{}) *gomock.Call {
+func (mr *MockCoreV1InterfaceMockRecorder) ServiceAccounts(arg0 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ServiceAccounts", reflect.TypeOf((*MockCoreV1Interface)(nil).ServiceAccounts), arg0)
 }
@@ -274,7 +279,7 @@ func (m *MockCoreV1Interface) Services(arg0 string) v12.ServiceInterface {
 }
 
 // Services indicates an expected call of Services.
-func (mr *MockCoreV1InterfaceMockRecorder) Services(arg0 interface{}) *gomock.Call {
+func (mr *MockCoreV1InterfaceMockRecorder) Services(arg0 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Services", reflect.TypeOf((*MockCoreV1Interface)(nil).Services), arg0)
 }
@@ -312,7 +317,7 @@ func (m *MockSecretInterface) Apply(arg0 context.Context, arg1 *v11.SecretApplyC
 }
 
 // Apply indicates an expected call of Apply.
-func (mr *MockSecretInterfaceMockRecorder) Apply(arg0, arg1, arg2 interface{}) *gomock.Call {
+func (mr *MockSecretInterfaceMockRecorder) Apply(arg0, arg1, arg2 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Apply", reflect.TypeOf((*MockSecretInterface)(nil).Apply), arg0, arg1, arg2)
 }
@@ -327,7 +332,7 @@ func (m *MockSecretInterface) Create(arg0 context.Context, arg1 *v1.Secret, arg2
 }
 
 // Create indicates an expected call of Create.
-func (mr *MockSecretInterfaceMockRecorder) Create(arg0, arg1, arg2 interface{}) *gomock.Call {
+func (mr *MockSecretInterfaceMockRecorder) Create(arg0, arg1, arg2 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Create", reflect.TypeOf((*MockSecretInterface)(nil).Create), arg0, arg1, arg2)
 }
@@ -341,7 +346,7 @@ func (m *MockSecretInterface) Delete(arg0 context.Context, arg1 string, arg2 v10
 }
 
 // Delete indicates an expected call of Delete.
-func (mr *MockSecretInterfaceMockRecorder) Delete(arg0, arg1, arg2 interface{}) *gomock.Call {
+func (mr *MockSecretInterfaceMockRecorder) Delete(arg0, arg1, arg2 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Delete", reflect.TypeOf((*MockSecretInterface)(nil).Delete), arg0, arg1, arg2)
 }
@@ -355,7 +360,7 @@ func (m *MockSecretInterface) DeleteCollection(arg0 context.Context, arg1 v10.De
 }
 
 // DeleteCollection indicates an expected call of DeleteCollection.
-func (mr *MockSecretInterfaceMockRecorder) DeleteCollection(arg0, arg1, arg2 interface{}) *gomock.Call {
+func (mr *MockSecretInterfaceMockRecorder) DeleteCollection(arg0, arg1, arg2 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "DeleteCollection", reflect.TypeOf((*MockSecretInterface)(nil).DeleteCollection), arg0, arg1, arg2)
 }
@@ -370,7 +375,7 @@ func (m *MockSecretInterface) Get(arg0 context.Context, arg1 string, arg2 v10.Ge
 }
 
 // Get indicates an expected call of Get.
-func (mr *MockSecretInterfaceMockRecorder) Get(arg0, arg1, arg2 interface{}) *gomock.Call {
+func (mr *MockSecretInterfaceMockRecorder) Get(arg0, arg1, arg2 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Get", reflect.TypeOf((*MockSecretInterface)(nil).Get), arg0, arg1, arg2)
 }
@@ -385,7 +390,7 @@ func (m *MockSecretInterface) List(arg0 context.Context, arg1 v10.ListOptions) (
 }
 
 // List indicates an expected call of List.
-func (mr *MockSecretInterfaceMockRecorder) List(arg0, arg1 interface{}) *gomock.Call {
+func (mr *MockSecretInterfaceMockRecorder) List(arg0, arg1 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "List", reflect.TypeOf((*MockSecretInterface)(nil).List), arg0, arg1)
 }
@@ -393,7 +398,7 @@ func (mr *MockSecretInterfaceMockRecorder) List(arg0, arg1 interface{}) *gomock.
 // Patch mocks base method.
 func (m *MockSecretInterface) Patch(arg0 context.Context, arg1 string, arg2 types.PatchType, arg3 []byte, arg4 v10.PatchOptions, arg5 ...string) (*v1.Secret, error) {
 	m.ctrl.T.Helper()
-	varargs := []interface{}{arg0, arg1, arg2, arg3, arg4}
+	varargs := []any{arg0, arg1, arg2, arg3, arg4}
 	for _, a := range arg5 {
 		varargs = append(varargs, a)
 	}
@@ -404,9 +409,9 @@ func (m *MockSecretInterface) Patch(arg0 context.Context, arg1 string, arg2 type
 }
 
 // Patch indicates an expected call of Patch.
-func (mr *MockSecretInterfaceMockRecorder) Patch(arg0, arg1, arg2, arg3, arg4 interface{}, arg5 ...interface{}) *gomock.Call {
+func (mr *MockSecretInterfaceMockRecorder) Patch(arg0, arg1, arg2, arg3, arg4 any, arg5 ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	varargs := append([]interface{}{arg0, arg1, arg2, arg3, arg4}, arg5...)
+	varargs := append([]any{arg0, arg1, arg2, arg3, arg4}, arg5...)
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Patch", reflect.TypeOf((*MockSecretInterface)(nil).Patch), varargs...)
 }
 
@@ -420,7 +425,7 @@ func (m *MockSecretInterface) Update(arg0 context.Context, arg1 *v1.Secret, arg2
 }
 
 // Update indicates an expected call of Update.
-func (mr *MockSecretInterfaceMockRecorder) Update(arg0, arg1, arg2 interface{}) *gomock.Call {
+func (mr *MockSecretInterfaceMockRecorder) Update(arg0, arg1, arg2 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Update", reflect.TypeOf((*MockSecretInterface)(nil).Update), arg0, arg1, arg2)
 }
@@ -435,7 +440,7 @@ func (m *MockSecretInterface) Watch(arg0 context.Context, arg1 v10.ListOptions) 
 }
 
 // Watch indicates an expected call of Watch.
-func (mr *MockSecretInterfaceMockRecorder) Watch(arg0, arg1 interface{}) *gomock.Call {
+func (mr *MockSecretInterfaceMockRecorder) Watch(arg0, arg1 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Watch", reflect.TypeOf((*MockSecretInterface)(nil).Watch), arg0, arg1)
 }
