@@ -1,5 +1,10 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: github.com/bakito/sealed-secrets-web/pkg/seal (interfaces: Sealer)
+//
+// Generated by this command:
+//
+//	mockgen -destination pkg/mocks/seal/mock.go --package seal github.com/bakito/sealed-secrets-web/pkg/seal Sealer
+//
 
 // Package seal is a generated GoMock package.
 package seal
@@ -10,7 +15,7 @@ import (
 	reflect "reflect"
 
 	seal "github.com/bakito/sealed-secrets-web/pkg/seal"
-	gomock "github.com/golang/mock/gomock"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // MockSealer is a mock of Sealer interface.
@@ -46,7 +51,7 @@ func (m *MockSealer) Certificate(arg0 context.Context) ([]byte, error) {
 }
 
 // Certificate indicates an expected call of Certificate.
-func (mr *MockSealerMockRecorder) Certificate(arg0 interface{}) *gomock.Call {
+func (mr *MockSealerMockRecorder) Certificate(arg0 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Certificate", reflect.TypeOf((*MockSealer)(nil).Certificate), arg0)
 }
@@ -61,7 +66,7 @@ func (m *MockSealer) Raw(arg0 seal.Raw) ([]byte, error) {
 }
 
 // Raw indicates an expected call of Raw.
-func (mr *MockSealerMockRecorder) Raw(arg0 interface{}) *gomock.Call {
+func (mr *MockSealerMockRecorder) Raw(arg0 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Raw", reflect.TypeOf((*MockSealer)(nil).Raw), arg0)
 }
@@ -76,7 +81,7 @@ func (m *MockSealer) Seal(arg0 string, arg1 io.Reader) ([]byte, error) {
 }
 
 // Seal indicates an expected call of Seal.
-func (mr *MockSealerMockRecorder) Seal(arg0, arg1 interface{}) *gomock.Call {
+func (mr *MockSealerMockRecorder) Seal(arg0, arg1 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Seal", reflect.TypeOf((*MockSealer)(nil).Seal), arg0, arg1)
 }
@@ -90,7 +95,7 @@ func (m *MockSealer) Validate(arg0 context.Context, arg1 io.Reader) error {
 }
 
 // Validate indicates an expected call of Validate.
-func (mr *MockSealerMockRecorder) Validate(arg0, arg1 interface{}) *gomock.Call {
+func (mr *MockSealerMockRecorder) Validate(arg0, arg1 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Validate", reflect.TypeOf((*MockSealer)(nil).Validate), arg0, arg1)
 }
