@@ -1,5 +1,10 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: github.com/bitnami-labs/sealed-secrets/pkg/client/clientset/versioned/typed/sealedsecrets/v1alpha1 (interfaces: BitnamiV1alpha1Interface,SealedSecretInterface)
+//
+// Generated by this command:
+//
+//	mockgen -destination pkg/mocks/ssclient/mock.go --package ssclient github.com/bitnami-labs/sealed-secrets/pkg/client/clientset/versioned/typed/sealedsecrets/v1alpha1 BitnamiV1alpha1Interface,SealedSecretInterface
+//
 
 // Package ssclient is a generated GoMock package.
 package ssclient
@@ -10,7 +15,7 @@ import (
 
 	v1alpha1 "github.com/bitnami-labs/sealed-secrets/pkg/apis/sealedsecrets/v1alpha1"
 	v1alpha10 "github.com/bitnami-labs/sealed-secrets/pkg/client/clientset/versioned/typed/sealedsecrets/v1alpha1"
-	gomock "github.com/golang/mock/gomock"
+	gomock "go.uber.org/mock/gomock"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	types "k8s.io/apimachinery/pkg/types"
 	watch "k8s.io/apimachinery/pkg/watch"
@@ -63,7 +68,7 @@ func (m *MockBitnamiV1alpha1Interface) SealedSecrets(arg0 string) v1alpha10.Seal
 }
 
 // SealedSecrets indicates an expected call of SealedSecrets.
-func (mr *MockBitnamiV1alpha1InterfaceMockRecorder) SealedSecrets(arg0 interface{}) *gomock.Call {
+func (mr *MockBitnamiV1alpha1InterfaceMockRecorder) SealedSecrets(arg0 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "SealedSecrets", reflect.TypeOf((*MockBitnamiV1alpha1Interface)(nil).SealedSecrets), arg0)
 }
@@ -101,7 +106,7 @@ func (m *MockSealedSecretInterface) Create(arg0 context.Context, arg1 *v1alpha1.
 }
 
 // Create indicates an expected call of Create.
-func (mr *MockSealedSecretInterfaceMockRecorder) Create(arg0, arg1, arg2 interface{}) *gomock.Call {
+func (mr *MockSealedSecretInterfaceMockRecorder) Create(arg0, arg1, arg2 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Create", reflect.TypeOf((*MockSealedSecretInterface)(nil).Create), arg0, arg1, arg2)
 }
@@ -115,7 +120,7 @@ func (m *MockSealedSecretInterface) Delete(arg0 context.Context, arg1 string, ar
 }
 
 // Delete indicates an expected call of Delete.
-func (mr *MockSealedSecretInterfaceMockRecorder) Delete(arg0, arg1, arg2 interface{}) *gomock.Call {
+func (mr *MockSealedSecretInterfaceMockRecorder) Delete(arg0, arg1, arg2 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Delete", reflect.TypeOf((*MockSealedSecretInterface)(nil).Delete), arg0, arg1, arg2)
 }
@@ -129,7 +134,7 @@ func (m *MockSealedSecretInterface) DeleteCollection(arg0 context.Context, arg1 
 }
 
 // DeleteCollection indicates an expected call of DeleteCollection.
-func (mr *MockSealedSecretInterfaceMockRecorder) DeleteCollection(arg0, arg1, arg2 interface{}) *gomock.Call {
+func (mr *MockSealedSecretInterfaceMockRecorder) DeleteCollection(arg0, arg1, arg2 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "DeleteCollection", reflect.TypeOf((*MockSealedSecretInterface)(nil).DeleteCollection), arg0, arg1, arg2)
 }
@@ -144,7 +149,7 @@ func (m *MockSealedSecretInterface) Get(arg0 context.Context, arg1 string, arg2 
 }
 
 // Get indicates an expected call of Get.
-func (mr *MockSealedSecretInterfaceMockRecorder) Get(arg0, arg1, arg2 interface{}) *gomock.Call {
+func (mr *MockSealedSecretInterfaceMockRecorder) Get(arg0, arg1, arg2 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Get", reflect.TypeOf((*MockSealedSecretInterface)(nil).Get), arg0, arg1, arg2)
 }
@@ -159,7 +164,7 @@ func (m *MockSealedSecretInterface) List(arg0 context.Context, arg1 v1.ListOptio
 }
 
 // List indicates an expected call of List.
-func (mr *MockSealedSecretInterfaceMockRecorder) List(arg0, arg1 interface{}) *gomock.Call {
+func (mr *MockSealedSecretInterfaceMockRecorder) List(arg0, arg1 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "List", reflect.TypeOf((*MockSealedSecretInterface)(nil).List), arg0, arg1)
 }
@@ -167,7 +172,7 @@ func (mr *MockSealedSecretInterfaceMockRecorder) List(arg0, arg1 interface{}) *g
 // Patch mocks base method.
 func (m *MockSealedSecretInterface) Patch(arg0 context.Context, arg1 string, arg2 types.PatchType, arg3 []byte, arg4 v1.PatchOptions, arg5 ...string) (*v1alpha1.SealedSecret, error) {
 	m.ctrl.T.Helper()
-	varargs := []interface{}{arg0, arg1, arg2, arg3, arg4}
+	varargs := []any{arg0, arg1, arg2, arg3, arg4}
 	for _, a := range arg5 {
 		varargs = append(varargs, a)
 	}
@@ -178,9 +183,9 @@ func (m *MockSealedSecretInterface) Patch(arg0 context.Context, arg1 string, arg
 }
 
 // Patch indicates an expected call of Patch.
-func (mr *MockSealedSecretInterfaceMockRecorder) Patch(arg0, arg1, arg2, arg3, arg4 interface{}, arg5 ...interface{}) *gomock.Call {
+func (mr *MockSealedSecretInterfaceMockRecorder) Patch(arg0, arg1, arg2, arg3, arg4 any, arg5 ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	varargs := append([]interface{}{arg0, arg1, arg2, arg3, arg4}, arg5...)
+	varargs := append([]any{arg0, arg1, arg2, arg3, arg4}, arg5...)
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Patch", reflect.TypeOf((*MockSealedSecretInterface)(nil).Patch), varargs...)
 }
 
@@ -194,7 +199,7 @@ func (m *MockSealedSecretInterface) Update(arg0 context.Context, arg1 *v1alpha1.
 }
 
 // Update indicates an expected call of Update.
-func (mr *MockSealedSecretInterfaceMockRecorder) Update(arg0, arg1, arg2 interface{}) *gomock.Call {
+func (mr *MockSealedSecretInterfaceMockRecorder) Update(arg0, arg1, arg2 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Update", reflect.TypeOf((*MockSealedSecretInterface)(nil).Update), arg0, arg1, arg2)
 }
@@ -209,7 +214,7 @@ func (m *MockSealedSecretInterface) UpdateStatus(arg0 context.Context, arg1 *v1a
 }
 
 // UpdateStatus indicates an expected call of UpdateStatus.
-func (mr *MockSealedSecretInterfaceMockRecorder) UpdateStatus(arg0, arg1, arg2 interface{}) *gomock.Call {
+func (mr *MockSealedSecretInterfaceMockRecorder) UpdateStatus(arg0, arg1, arg2 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "UpdateStatus", reflect.TypeOf((*MockSealedSecretInterface)(nil).UpdateStatus), arg0, arg1, arg2)
 }
@@ -224,7 +229,7 @@ func (m *MockSealedSecretInterface) Watch(arg0 context.Context, arg1 v1.ListOpti
 }
 
 // Watch indicates an expected call of Watch.
-func (mr *MockSealedSecretInterfaceMockRecorder) Watch(arg0, arg1 interface{}) *gomock.Call {
+func (mr *MockSealedSecretInterfaceMockRecorder) Watch(arg0, arg1 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Watch", reflect.TypeOf((*MockSealedSecretInterface)(nil).Watch), arg0, arg1)
 }
